requirements.txt
bcrypt==3.1.4
bottle==0.11.7
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.13.5
passlib==1.7.1
peewee==2.10.1
pudb==2018.1
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.2
six==1.11.0
urwid==2.0.1
Werkzeug==0.14.1
WTForms==2.2.1
